requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python-headless # <--- CHANGED THIS LINE
pywavelets
pyzbar
qrcode
